locsAppBack/requirements.txt
defusedxml==0.4.1
Django==1.8.4
django-allauth==0.23.0
django-cors-headers==1.1.0
django-filter==0.11.0
django-rest-auth==0.5.0
djangorestframework==3.2.4
Markdown==2.6.2
mongoengine==0.10.0
oauthlib==1.0.3
psycopg2==2.6.1
pymongo==3.0.3
python3-openid==3.0.7
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
wheel==0.24.0
gunicorn
